Create README.md
I did my Ph.D in Genetics (Evolution and Development) at the Max Planck Institute for Developmental Biology (Tuebingen, Germany) with bioinformatics training as an important part of it. After graduation, my research was mainly in the field of Immunology. I switched to the field of Microbiology and Virology a few years ago.